Revert changes that were merged incorrectly
